separate scanner tests from scanner service
