+++ src/scanner.rs
@@ -3,3 +3,6 @@ pub mod token;
 
 pub use scanner::Scanner;
 pub use token::Token;
+
+#[cfg(test)]
+mod tests;
